Moved Cohere Transcribe into a file of its own
This model has earned a file. It is the most accurate open recognizer there is
and it has opinions — about how much audio it wants at a time, about what
happens when it is handed a stretch with no speech in it, about what it says
when it would rather say nothing — and each of those was found the hard way and
each has its reason written beside it. That is a lot to keep threaded through a
file about captions in general.

So cohere.go holds the model and everything it asks for, and the day it is
replaced that file is deleted. Two references break when it goes, its line in
the catalog and its case in quirksFor, and the compiler names both by line
number. That is a better guarantee than a comment asking somebody to remember,
and I checked it: delete the file and the build says exactly those two things
and nothing else.

Nothing else reaches into it. The recognizer, the caption encoders, the
injector and the tune gate still have no idea which model they are serving,
which is the property worth keeping and the reason this was worth doing rather
than simply tidy.
